+++ server/src/test/java/dpas/server/service/ReliableByzantineServicePostTest.java
@@ -64,7 +64,7 @@ public class ReliableByzantineServicePostTest {
 
     private static Contract.Announcement[] _requests;
 
-    private static long _seq = 0;
+    private static int _seq = 0;
 
     private static final String MESSAGE = "MESSAGE";
