+++ server/src/test/java/dpas/server/service/ReliableServerPostGeneralTest.java
@@ -59,7 +59,7 @@ public class ReliableServerPostGeneralTest {
 
     private static Contract.Announcement[] _requests;
 
-    private static long _seq = 0;
+    private static int _seq = 0;
 
     private static final String MESSAGE = "MESSAGE";
